Move automated season code to a separate module
