+++ app/services/__init__.py
@@ -1,10 +1,8 @@
 import logging
-from datetime import timedelta
 
 from django.conf import settings
-from django.utils import timezone
 
-from app import models, serializers, tasks
+from app import models, tasks
 from app.services.ratings import update_ratings_from_match
 from app.services.trophy import create_trophies
 
@@ -62,58 +60,6 @@ def update_tournament_state(tournament):
         create_trophies(tournament)
 
 
-def update_seasons_state():
-    for season in models.Season.objects.all():
-        update_season_state(season)
-
-
-def update_season_state(season):
-    now = timezone.now()
-
-    if season.end_date < now:
-        season.active = False
-        season.save(update_fields=["active"])
-
-
-def create_automated_seasons():
-    now = timezone.now()
-    start_date = now.replace(hour=0, minute=0, second=0, microsecond=0)
-    end_date = now.replace(
-        hour=23, minute=59, second=59, microsecond=999999
-    ) + timedelta(days=1)
-    last_automated_season = (
-        models.Season.objects.filter(is_automated=True)
-        .order_by("-automated_number")
-        .first()
-    )
-
-    if last_automated_season:
-        next_number = last_automated_season.automated_number + 1
-    else:
-        next_number = 1
-
-    if last_automated_season and last_automated_season.active:
-        return {"status": "active season already exists"}
-
-    name = f"Automated Season {next_number}"
-
-    data = {
-        "name": name,
-        "is_automated": True,
-        "automated_number": next_number,
-        "start_date": start_date,
-        "end_date": end_date,
-    }
-    serializer = serializers.SeasonSerializer(data=data)
-    if serializer.is_valid():
-        logger.info(f'creating automated season "{name}"')
-        serializer.save()
-
-        create_ratings_for_season(serializer.instance)
-    else:
-        logger.warning(f'failed to create season "{name}"')
-
-
 def recalculate_ratings_for_season(season):
     updated_count = 0
     matches_to_update = season.matches.filter(ran=True).count()
@@ -144,8 +90,3 @@ def recalculate_ratings_for_season(season):
             )
 
     logger.info(f"Finished recalculating rankings for season '{season.name}'")
-
-
-def create_ratings_for_season(season):
-    for agent in models.Agent.objects.active().all():
-        models.AgentRatings.objects.create(season=season, agent=agent, game=agent.game)
